restructure codebase.
